Move settings to other location
Settings have been moved to a location under application data.
Depending on the OS, the location of the settings file is now
one of the following.

* Windows: ~/AppData/Local/rectball
  For instance: C:\Users\danirod\AppData\Local\rectball
* Linux: ~/.rectball
  For instance: /home/danirod/.rectball
* MacOS X: ~/Library/Application Support/rectball
  For instance: /Users/danirod/Library/Application Support/rectball

As of this commit this has been tested on MacOS X. Before merging
into upstream I will check that this works as intended on Linux and
Windows too and that this doesn't introduce regression bugs on
Android.