Move the state of the game to GameState
GameState will contain information about the current score and
time in a numeric value. The class is part of the model so
the game can use it to update the views.

It is expected that this class will contain in the near future
information about the board, to separate the board model and
logic from view representation.