Create better timer graphics
The old timer texture has been replaced by a new and more stylish
texture. Also, the timer texture is no more a texture region. Now
the timer texture is a single light gray texture. To draw the
timer green, tint the batch to green. To draw the timer red, tint
the batch to red.

The AssetManager has been updated and the previous timer has been
completely removed from code and from the skin file as well.