Call setUpInterface() during show()
Instead of composing the layout on load() method, now the layout
is composed on show() method. This means that every time that the
screen is switched, the layout for the target screen is revalidated.

This has been done because some screens have to change the information
they display every time the screen is displayed and this required
unnecesary code that could be removed in setUpLayout() were called
on show().

Now, every time the screen is displayed, the layout can be created
with the new values for the screen.