Set up the game in setUpLayout()
Now that setUpLayout() is invoked on every screen show, a cleanup
can be made. New actors can be created when the game is about to
start instead of reusing old actors. This removes code because
now there is no need for resetting the state of the actors, just
create new ones.