Enable cancelling screen jumping events
Previously when a TextButton was pressed, the button was unchecked.
This was introduced in PR #17 and made the buttons not appear
clicked once pressed.

However, this has introduced an interesting bug. When you uncheck
the button, the changed event is fired again, making the game
process the event twice. Consequences:

* The screen is added to the stack twice, so the user has to press
  Back button twice.
* The sound was played twice.

Instead of unchecking the button, apparently you can just cancel
the event to make the button not change the state. This will not
change the checked status of the button and this only fires one
single event.