Style the dialog
Old dialog was so ugly and clunky. It was grayish, sad and the
buttons didn't look good. So I made a new style for the dialog
and kinda worked on the dialog API.

Since both dialogs, Confirm and Message, should act similar, I
made a Common dialog to apply changes to both dialogs at the
same time. Working with Dialog class in libGDX is a nightmare
because you can't easily set the size for things, but at least
for now:

* The label should be displayed with the maximum size possible.
* Buttons should try to expand to fill the width of the dialog.
  Confirm buttons will be half the size, message dialog will
  be the total size.