Add strings to es_AR locale
I'm not pretty sure on what does libGDX when it doesn't find a
string in a locale file. If a string is not defined in es_AR,
I don't know if it fallbacks to es because of the same I18N
or if it fallbacks to the generic one (English). Just in case
I add these lines.