Add platforms support
Because libGDX is a multiplatform library, multiplatform code
should be handled in a special way. The most common way for handling
code that is supposed to run in particular platforms, is to add
the code to the platform project and pass it to the core game
as a parameter. They will need to implement a particular interface.