Enable sharing screenshots on desktop
A new service has been added to the platform API. Now the game
can request to share a screenshot from the game. At the moment this
is only enabled for desktop.

Client can generate a screenshot and then use the sharing API to
share the screenshot. Each platform might implement this on a
different way. Desktop is set to save the photo on the hard drive
after asking the user for a valid location. Android might decide
to show a sharing intent.