Make the game actually be fullscreen
Instead of using a fixed viewport width and height, use a fixed
width and attempt to guess the height coming from the aspect
ratio.

This removes any border that appears on the game when it runs on
a device that doesn't have 16:10 aspect ratio such as 4:3 or 16:9.
Still has to be tested that every screen works, specially on 4:3,
which is the smallest height.