Fixes memory leak on TextureRegionDrawable
Previously new TextureRegionDrawables were being generated every
frame. This was partially doing a hog on the resource consumptions
for the game and eventually causing memory leaks.

The proposed solution is not the most beautiful one but it works.
I create the TextureRegionDrawables when the ball actor is created
and the same drawables are used instead of creating new ones.