Replace CombinationFinder by Singleton
CombinationFinder seems to be one of the weightest classes being
used in Rectball. It is instantiated a lot and it creates lots of
things. It is a hog on the garbage collection. Therefore, create
a Singleton so that only one instance can be created.

Additionally, most of the ArrayLists that were using in the
calculateCombinations() method are gone and instead both loops
have been inlined in the same function. It works as per unit tests.