Save state on Android Bundles
Introduce the concept of savedState. Android allows you to save the
state of an activity so that if the activity is destroyed because
of the system needing that resources, it can then recreate the
activity.

This is NOT a persistent save, it's just a temporal save in case the
activity is destroyed because of low resources, and it's exactly
what I need to avoid losing Rectball state in case the player
minimizes the game for such a long time that the system kills the
game.

What this change introduces is the packing of all the state-related
variables into the GameState variable. Then, a system is provided
for storing that GameState into a serialized string when the state
has to be saved.

In case Android kills the activity, the onCreate method will
re-create the activity with all the previous state and the game will
continue thanks to the changes introduced in the GameScreen that check
whether a previous state exists and recreates it.