Add a new loading screen
When the game is restoring from a destroyed state in a game screen,
a different loading screen will be used to notify the user that the
game is restoring from the saved state.

To not alarm the user, the loading screen will "recompose" the balls
as the assets are loading in the same position where they should be
during normal game.