Remove @2x assets
These assets only work for iOS and Mac OS X, which are not the
target platform for this game (Android). Desktop support is just
a side effect of the current development workflow.

Therefore, I'm removing the @2x assets to cut up to half a megabyte
of space, which comes good in terms of reducing the application
size.

I'm keeping the useHDPI flag enabled because that seems to make
the application look good on OS X despite not having @2x assets.