Upgrade Gradle to the latest version
It has been a while since this project is not updated and Gradle
has been updated several times. Upgraded from Gradle 2.4 to
Gradle 2.11 for the Wrapper.

Plus, Android Build Tools have been updated too so they have been
upgraded from 1.2.3 to 1.5.0 in order to work with the latest
versions of Gradle.