Add Share button to statistics
The user can share screenshots of the Statistics screen by pressing
the Share button, now located at the Statistics button. The screenshot
is resized to match the size of the widget, so no scroll bars are
visible.

Screenshot is taking by generating a Frame Buffer Object and rendering
the contents of the table widget to that Frame Buffer Object, then
converting the FBO to a Pixmap and then saving the pixmap via the
standard Sharing system.

This means that on desktop the user will see the Save as... window,
while on Android an Intent for sharing the photo will be triggered
letting the user share the screenshot via other apps.

Closes #50.