Remove support for desktop
Sorry for people playing still on desktop, but I want to focus on
providing a better experience on mobile phones. So, no more Rectball for
desktop. Of course you can still play Rectball 0.4.7 but that version
won't probably be maintained anymore by the time the next patch or
version goes out.