Count how many times a hint is requested
A new statistic is added for tracking the number of times a hint is
requested by the user. This counter is incremented whenever a
combination has been found but the game reveals the Hint button was
pressed. If the game is over but a hint has been requested, it is
counted as well because the user pressed the button anyway.

This was requested in #70.