Split the project in Open Source vs Google Play
Two flavours of the project are created: GPE and OSP. Both are open
source now, but GPE is a special version linked against the Google Play
SDK and the Firebase SDK that is able to report analytics to Google
Analytics and game data to Google Play.

While OSP is just the regular open source version that doesn't contain
non-free code. Some folks may prefer to use this version, which can just
be compiled straight from the code without providing any API keys or
accepting agreements. If I ever send Rectball to F-Droid, the OSP is the
way to go.