Move GPE specific items from AndroidManifest
I did a cold build of the OSP version removing any GPE files and I
discovered that the thing was not compiling because it expected to have
a few config files from the GPE version.

Those parts of the Manifest that depend on GPE classes and libraries
have been moved into a separate Manifest that is on the GPE project and
merged wit the main Manifest when the project gets built.

Additionally a dummy google-services.json is generated in app/ because
the project doesn't seem to build without it as the Google GMS plugin is
missing a few configuration files.