Remove Stats and Statistics from the Platform
They were in the platform because each platform stored them on a
different location. This is not the case anymore as desktop was removed
a long time ago, so it makes no sense to delegate storing these files
into the platform.

This commit renames LegacyScores and LegacyStatistics as ScoresIO and
StatisticsIO, and then removes the connection with the Platform. Scores
and Statistics interfaces are also dropped. The new Scores and
Statistics are added to the RectballGame module.