Disables ProGuard optimizations in Statistics
Apparently I've been not properly saving my statistics since I started
using ProGuard in may. ProGuard optimizes the names for the fields in
Statistics classes so that instead of "color", "total", "sizes" the
fields are renamed to "a", "b" and "c". Unfortunately, that also affects
the generated JSON through serialization and reflection. And,
unfortunatelly as well, I'm hardcoding the expected strings to read as
the non-optimized values. Therefore, they are never properly read
because the JSON contains an "a" where my logic expects a "color".