refactor: extract ball selection logic
This commit will create a class called ClassicBoardSelectionListener
containing the input listener used to select and unselect balls, taking
it out of the BallActor and associating it to the BoardActor.

The goal of the change is to make easier in the future to add a different
listener that behaves using drag movements instead of taps, which will
require to be added at the Board level rather than at the Ball level, to
support gestures that span multiple balls.