feat: add drag selection method
This commit adds a new method for selecting combinations that is based
around the idea of dragging one corner against its diagonal one (i.e.
picking the ball in the NW corner and dragging up to the SE corner, or
picking the ball in the NE corner and dragging up to the SW corner, on
either direction).