refactor: cleanup BallActor and BoardActor
In order to refactor the animation in the LoadingScreen, I need to
cleanup the entities that BallActor depend upon, such as BoardActor.

To remove this dependency, instead of testing whether the board is
coloured on each frame, the setColoured message sent to the board is
propagated to the BallActor children instances, so that they can be
declared as coloured or not depending on the status of the board itself.

Some dead code that prevented this change to compile successfully was
removed too.