feat: rewrite the tutorial and update the texts
The tutorial had to be updated because the board is now dragged rather
than touched. At the same time, I took the opportunity to rewrite it
in Kotlin. I believe the code is more concise and the state machine is
gone and replaced with a few modals with callbacks, easier to
understand. In any case, I'll know if this was a good refactor or not
in a few years. (Dear future Dani, I'm sorry!)