fix: don't enable inmersive mode by default
libGDX 1.12 makes inmersive mode the default, but this interferes with
my code to enable or disable the fullscreen mode depending on what the
user has set in the settings.

Disable it.

As a sidenote: should fix the deprecation issues and change to
WindowInsetsController, not just for fullscreen mode, but also to fix
the black bars when the device has a non-rectangular shape.