feat: wakelock only during the game screen
This commit will remove the wakelock request that happens whenever the
application starts. Instead, when the GameScreen gets presented, the
application will request a wakelock.

This has the consequence that leaving the phone on in the main menu
screen (or any other screen) will not keep the screen on all the times.
However, when a game starts, the screen is kept on in order to allow the
player to play.

Pausing the game, such as pressing the X button or switching to a
different application, will deactivate the wakelock. When the game is
paused, the wakelock should be dropped either way. The wakelock is also
dropped once the game finishes.