fix: WindowInsets not being configured on Android 4.4
Android 4.4 is this weird middlepoint where the game still presents
using translucent status bars (so there is content behind the status
bar), but the Compat code to set the insets does not work anymore
because AndroidX does not support Android below 5 anymore.

In this commit I use the strategy pattern to provide a fallback relying
on some really deprecated APIs if the game is running under Android 4.4,
in order to set the insets by querying information about the system.

Either way, I should test this on a x86 Android 4.4 VM later. Android
4.4 support is going to be removed later, but I still want to finish the
game before doing that.