refactor: remove fullscreen mode
In this commit, everything related to the classic fullscreen mode is
removed. No more checkbox, no more integration code to turn on the
fullscreen, no more quit button in the homescreen, no screens.

As covered by the roadmap, fullscreen mode is completely broken on
recent Android versions and it is just easier to remove it than to fix
it, since modern phones have a different way to present things where the
rationale behind fullscreen mode is not so obvious.