refactor: migrate gems to Skin Composer
An interesting commit where I start using Skin Composer to manage
the graphical assets of the game, and to keep track of the original
SVG files in the GitHub repository too, to avoid having to look
for three hard drives whenever I want to open them.

I start by tracking the color gems. In this commit the following
changes are done to the assets:

* Imports the gems SVG file into skin/svg/gems.svg.
* Renders some high resolution gems into skin/images/gem_*.png.
* Creates a Skin Composer project into skin/scmp/rectball.scmp.
* Exports the future skin, which at the moment only contains the
  gems, into app/src/main/assets/skin/*.

Then changes are made in code so that the gems are loaded from
the new skin rather than from the texture file. This requires to
do some specific changes:

* Since the Skin file is loaded as part of the AssetManager, it is
  required that the Skin file is loaded even before presenting the
  loading screen for the rest of the assets.
* I believe that in the future, the loading screens may even go
  away, as probably the skin file may swallow most other files
  and pregenerated resources which are currently loaded in the
  loading screens.

To assert that the new files are being used, the legacy board assets
in app/src/main/assets/board/*.png have been removed, both from the
repository assets folder, and from the AssetManager itself.