feat: new skin engine
This commit replaces most of the old RectballSkin class by an actual
libGDX skin file, based on the work done in previous commits. Because
there are many visual changes that impact the game, this commit also
revamps notable parts of the user interface.

This is not the definitive new look and feel, because this will change
in the future, but this is a work in progress that hasn't finished and
that is due for v0.7.

Some improtant changes in this version:

* Most of the table components (such as the settings screen or the main
  menu) have been refactored now that I know more about Scene2D's Table
  layouts.

* Screens that present non-game menus, such as settings, about or
  settings, now extend a new abstract screen called MenuScreen. The
  common features: a back button on top, and a scrollpane in case the
  elements are larger than the screen.

* The game screen HUD now doesn't make use of yellow borders. This was
  done because the new buttons make use of ninepatches and they just
  don't fit anymore if I keep using yellow borders.

* The dialogs now have different animations, in an attempt to avoid
  using alpha effects, since the nine patchs are slightly buggy when
  alpha is not 0 or 1.