feat: add vibration and wakelock control
Haptics is a class that allows to control things such as vibration. It
is bound to RectballGame, so methods can be accessed by using
game.haptics or game.getHaptics().

Sidenote: I've tried to use VibrationType.LIGHT, but it turns out that
this is only supported starting with Android O. Maybe I could detect the
Android version and use a fallback if the version is older than Android
O, but not at this time.

Also, this commit introduces the AppSettings class, which is a way to
centralize all the queries to the settings. This is going to be useful
once I go back to GdxPreferences instead of relying directly on Android
SharedPreferences when adding PC support again.

As a proof of concept, I allow controlling the wakelock. The methods for
requesting the wakelock have been updated properly according to this.