refactor: simplify back / escape button logic
Instead of having a separate constructor and having to manually set the
InputProcessor on the screens, there is now an overridable method called
`escape` in AbstractScreen with the code that should be triggered if
Input.Keys.BACK or Input.Keys.ESCAPE are pressed.

The second constructor is removed, and the InputMultiplexer is always
instantiated to avoid messing with AbstractScreen's stage, which should
always be an input processor anyway.

The `escape` method is successfully overriden in the following cases:

* GameScreen: rather than leaving, should ask the user first.
* GameOverScreen: clear the entire stack, go back to the start menu.
* Loading screens: should not do anything, since these screens cannot be
  stopped.

For any other case, the default implementation pops the current screen
from the screen stack and plays a default sound, which is the same that
it did before.

Also, the InputAdapter for the escape behaviour is now a non-static
nested class rather than a static nested class, since this is something
that I did not know about Java back in the days.