refactor: lazy load screens
In this commit, the StateBasedGame class will stop preloading screens
and binding each screen with an ID number. Instead, the stack inside the
StateBasedGame now expects instances of Screen to be provided, so
calling pushScreen() requires passing a Screen instance.

Therefore, all numerical identifiers are removed, and whenever a screen
needs to be presented, it will be instantiated on demand. Note that I
haven't checked whether the current dispose() strategy is correct or if
resources are not being cleaned yet.

This refactor was intensively chased because now I can start assuming
that libGDX is fully initialised in each screen constructor and thus
stop doing lazyinit to everything.  More refactors in the future will
start initialising stuff in the constructor.