refactor: migrate from Android strings to GDX I18nBundle
One more step towards making the game portable again. In this commit
the strings.xml files are replaced by I18nBundles. A Locale class is
used because I know how things are going to end, plus depending on the
flavor in use, I need to get the strings either from bundles/strings
or from bundles/google_play (the later one should not be commited when
sending the game to F-Droid, for instance.)