fix: MigratorV1 breaks if V1 files are missing
If the V1 SharedPreferences file exists but the V1 scores or stats
files do not, it crashes. This is because the exists() method checks
that either any of them exists, but then tries to access all of
them regardless of whether they exist or not.

After this fix, migrating a V1 file will only migrate the scores and
statistics if those file exists, allowing cases such as having only
1 or 2 of the 3 files available.