refactor: enhanced versioning infrastructure
Instead of three variables for setting the game version, there are only
two now, and one of them is optional.

The main identifier is versionFamily, which is usually left alone in
the gradle.properties file, and is only bumped whenever a new stable
version is released, so it can be done manually.

You can set the buildNumber manually using the Gradle variable and it
will be respected, but it is not mandatory. The main version that is
presented inside the app will only be the versionFamily if there is no
build number.

If the buildNumber is set to something, the version code presented
inside the game will have the build number appended at the end.

The build number can now just be overridden by the GITHUB_RUN_ID and
GITHUB_RUN_OFFSET environment variables, in order to use GitHub Actions
as the source of truth, without requiring to use outputs.

The Gradle script should detect Git repositories and use the version
infered from the Git tag that points to the current HEAD, as long as the
current HEAD has been tagged to something like v.X.Y.ZZZ.

Since the information to create version numbers is now derived with less
command line arguments, the GitHub Actions workflow file has been
modified so that it doesn't accept the parameters from command line, but
rather it uses the versionFamily in the gradle.properties file and the
GITHUB_* environment variables to craft the build number.